display: Automatically adjust the base scale if the combination of individual monitor scales is such that it makes sense to.
For example, you have 2 monitors:

1x    2x      (base scale 2)

You change the second one to 1x:

1x    1x      (base scale 2)

It makes no sense to still have a base scale of 2, so we should
switch it to 1 on the user's behalf.

The same should happen in the other direction (2 monitors at 2x
should auto-set a base scale of 2).